Bug Fix: Commit MUST fail if any operation fails.
Prior to this change, the commit() function used to do all the
operations, even if one of them failed. In scenarious like a file is
deleted first and then tried to share will throw errors on the second
operation. The commit was doing both the operations and returning
success for 1st and failure for second. If there were any third
operation, it would do it. This is not desirable because if in a commit
an error occurs, we would like to cancel at the point where failure
happened. This would leave the system in a dangling state because the
commit had failed. However, since all of this would happen in its own
separate version, developer can remove that whole version entirely.

With this change, the commit will fail and stop as soon as it hits an
error. Also the final result of commit will say "failure" because it was
not completed.