Random Property Support
This project now supports for random user data that user might want to
store. If the property is not a mandatory property, the name/value gets
copied to the appropriate node. If a change metadata needs to be send,
it needs to be send inside a JSON Object whose key must be
"change_metadata".

E.g.
{
    "node_version": {
        "nodeId": "1",
        "property1": "test",
        "change_metadata": {
            "change_prop": "test change 1",
            "change_prop2": "test change 2"
        }
    }
}

In above, property1 gets copied to the node itself and keys inside
change_metadata gets copied to the relationship.