REST POST Mandatory Property Data Guideline
The Mandatory Property POST data was previously not strignent to follow
a case. This was an inconsistent behaviour with optional properties and
was causing other troubles and confusion.

The mandatory properties are now defined as camel case and the user MUST
match the property key (case sensitive) to use that property.